requirements.txt
python-docx>=2.8.1
python-dotenv>=0.19.0
openai>=1.0.0
google-generativeai>=0.3.0
keyboard>=0.13.5